[CI&hub] TorchHub integration tests (#362)
