+++ hubconf.py
@@ -1,6 +1,6 @@
 from asteroid import models
 
-dependencies = ["torch"]
+dependencies = ["torch", "numpy", "scipy", "asteroid_filterbanks"]
 
 
 def demask(name_url_or_file=None, **kwargs):
